player contains a buffer and broke stuff up
