+++ crates/bottomless-pit/src/player/mod.rs
@@ -1,4 +1,2 @@
-use super::settings::Inputs;
-
 mod attack;
 pub mod player;
